chore: reorganize the codebase for new env concept
