+++ pkg/cluster/addons/metallb/metallb_test.go
@@ -1,4 +1,4 @@
-package metallb_test
+package metallb
 
 import (
 	"net"
@@ -6,8 +6,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	ktfmetal "github.com/kong/kubernetes-testing-framework/pkg/metallb"
-	ktfnet "github.com/kong/kubernetes-testing-framework/pkg/networking"
+	ktfnet "github.com/kong/kubernetes-testing-framework/pkg/util/networking"
 )
 
 func TestHelperFunctions(t *testing.T) {
@@ -15,7 +14,7 @@ func TestHelperFunctions(t *testing.T) {
 		IP:   net.IPv4(192, 168, 1, 0),
 		Mask: net.IPv4Mask(0, 0, 0, 255),
 	}
-	ip1, ip2 := ktfmetal.GetIPRangeForMetallb(network)
+	ip1, ip2 := getIPRangeForMetallb(network)
 	assert.Equal(t, ip1.String(), net.IPv4(192, 168, 1, 240).String())
 	assert.Equal(t, ip2.String(), net.IPv4(192, 168, 1, 250).String())
 	assert.Equal(t, ktfnet.GetIPRangeStr(ip1, ip2), "192.168.1.240-192.168.1.250")
