+++ pkg/cluster/kind/client.go
@@ -11,28 +11,6 @@ import (
 )
 
 // -----------------------------------------------------------------------------
-// Public Functions - Cluster Management
-// -----------------------------------------------------------------------------
-
-// CreateCluster creates a new cluster using Kubernetes in Docker (KIND).
-func CreateCluster(name string) error {
-	// TODO: for now using CLI and outputting to stdout/stderr
-	// later we should switch to using the libs.
-	cmd := exec.Command("kind", "create", "cluster", "--name", name)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	return cmd.Run()
-}
-
-// DeleteKindCluster deletes an existing KIND cluster.
-func DeleteKindCluster(name string) error {
-	cmd := exec.Command("kind", "delete", "cluster", "--name", name)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	return cmd.Run()
-}
-
-// -----------------------------------------------------------------------------
 // Public Functions - Helper
 // -----------------------------------------------------------------------------
