+++ pkg/cluster/kind/kong_proxy_readiness.go
